Extract Components(CustomVideoPlayer, ScrollLink) & Hooks(useScrollToHash)
