+++ src/assets/styles/sections.css
@@ -79,172 +79,6 @@
     justify-content: center;
     margin-bottom: 0.5rem;
 }
-#root .hero-section .video-wrapper {
-    border-radius: 1rem;
-    overflow: hidden;
-    box-shadow: 0 12px 40px rgba(0, 0, 0, 0.7);
-    max-width: 100%;
-}
-#root .hero-section .video-wrapper video {
-    display: block;
-}
-#root .hero-section .video-overlay-text {
-    font-size: 1.1rem;
-    margin-top: 0.5rem;
-    text-align: center;
-}
-#root .hero-section .video-wrapper:hover {
-    transform: scale(1.01);
-    transition: transform 0.3s ease-in-out;
-}
-#root .video-wrapper{
-    position: relative;
-}
-#root .play-button {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    font-size: 4rem;
-    color: white;
-    background: rgba(0, 0, 0, 0.5);
-    border-radius: 50%;
-    width: 80px;
-    height: 80px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    z-index: 10003;
-    transition: transform 0.3s ease, opacity 0.3s ease;
-    cursor: pointer;
-}
-.custom-video-controls-wrapper {
-    position: absolute;
-    top: 47%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    z-index: 10003;
-    transition: opacity 0.3s ease, visibility 0.3s ease;
-    display: flex;
-    gap: 20px;
-}
-.custom-video-controls {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    width: 80px;
-    height: 80px;
-    background: rgba(0, 0, 0, 0.5);
-    border-radius: 50%;
-    cursor: pointer;
-    box-shadow: 0 6px 20px rgba(0, 0, 0, 0.4);
-    transition: transform 0.3s ease, opacity 0.3s ease;
-}
-.custom-video-controls:hover {
-    transform: scale(1.1);
-}
-.custom-video-controls svg {
-  pointer-events: none;
-  width: 60px;
-  height: 60px;
-}
-.custom-video-controls-wrapper.hidden {
-  opacity: 0;
-  visibility: hidden;
-}
-.pause-icon,
-.play-icon {
-    opacity: 1;
-    visibility: visible;
-    transition: opacity 0.3s ease, visibility 0.3s ease;
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-}
-.custom-progress-bar {
-    width: 100%;
-    height: 8px;
-    background: rgba(255, 255, 255, 0.2);
-    border-radius: 4px;
-    cursor: pointer;
-    margin-top: 10px;
-    position: relative;
-    overflow: hidden;
-}
-
-.video-time-container {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-  margin-top: 10px;
-  user-select: none;
-}
-
-.time-label {
-  font-size: 0.9rem;
-  color: #fff;
-  min-width: 40px;
-  text-align: center;
-}
-
-.custom-progress-bar {
-  flex-grow: 1;
-  height: 8px;
-  background: rgba(255, 255, 255, 0.2);
-  border-radius: 4px;
-  cursor: pointer;
-  position: relative;
-  overflow: hidden;
-}
-
-.progress-filled {
-  height: 100%;
-  background: #FD8804;
-  transition: width 0.1s linear;
-}
-
-.progress-thumb {
-  position: absolute;
-  top: 50%;
-  transform: translate(-50%, -50%);
-  width: 14px;
-  height: 14px;
-  background: white;
-  border-radius: 50%;
-  box-shadow: 0 0 4px rgba(0,0,0,0.5);
-  pointer-events: none;
-}
-
-.custom-video-controls.mute-controls {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 32px; /* Smaller size to match time label height */
-  height: 32px;
-  background: rgba(0, 0, 0, 0.5);
-  border-radius: 50%;
-  cursor: pointer;
-  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.3);
-  transition: transform 0.3s ease, opacity 0.3s ease;
-}
-
-.custom-video-controls.mute-controls:hover {
-  transform: scale(1.1);
-}
-
-.custom-video-controls.mute-controls svg {
-  pointer-events: none;
-  width: 24px; /* Smaller icon to fit compact button */
-  height: 24px;
-}
-
-.mute-icon,
-.unmute-icon {
-  opacity: 1;
-  visibility: visible;
-  transition: opacity 0.3s ease, visibility 0.3s ease;
-}
 
 @media (min-width: 768px) {
     #root .hero-section {
@@ -1068,14 +902,171 @@
     }
 }
 
-/* Custom properties (ensure in :root in base.css) */
-:root {
-    --bgCust2: rgb(229, 239, 255);
-    --bgCust3: rgb(187, 212, 245);
-    --coCust6: rgb(74, 85, 104);
-    --bgCust4: rgb(2, 45, 94);
-    --coCust5: rgb(251, 250, 251);
-    --coCust7: rgb(127, 141, 165);
-    --coCust8: rgb(19, 30, 48);
+/* Custom Video Wrapper */
+#root .video-wrapper {
+    border-radius: 1rem;
+    overflow: hidden;
+    box-shadow: 0 12px 40px rgba(0, 0, 0, 0.7);
+    max-width: 100%;
+}
+#root .video-wrapper video {
+    display: block;
+}
+#root .video-overlay-text {
+    font-size: 1.1rem;
+    margin-top: 0.5rem;
+    text-align: center;
+}
+#root .video-wrapper:hover {
+    transform: scale(1.01);
+    transition: transform 0.3s ease-in-out;
+}
+#root .video-wrapper{
+    position: relative;
+}
+#root .play-button {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    font-size: 4rem;
+    color: white;
+    background: rgba(0, 0, 0, 0.5);
+    border-radius: 50%;
+    width: 80px;
+    height: 80px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    z-index: 10003;
+    transition: transform 0.3s ease, opacity 0.3s ease;
+    cursor: pointer;
+}
+.custom-video-controls-wrapper {
+    position: absolute;
+    top: 47%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    z-index: 10003;
+    transition: opacity 0.3s ease, visibility 0.3s ease;
+    display: flex;
+    gap: 20px;
+}
+.custom-video-controls {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 80px;
+    height: 80px;
+    background: rgba(0, 0, 0, 0.5);
+    border-radius: 50%;
+    cursor: pointer;
+    box-shadow: 0 6px 20px rgba(0, 0, 0, 0.4);
+    transition: transform 0.3s ease, opacity 0.3s ease;
+}
+.custom-video-controls:hover {
+    transform: scale(1.1);
+}
+.custom-video-controls svg {
+  pointer-events: none;
+  width: 60px;
+  height: 60px;
+}
+.custom-video-controls-wrapper.hidden {
+  opacity: 0;
+  visibility: hidden;
+}
+.pause-icon,
+.play-icon {
+    opacity: 1;
+    visibility: visible;
+    transition: opacity 0.3s ease, visibility 0.3s ease;
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+}
+.custom-progress-bar {
+    width: 100%;
+    height: 8px;
+    background: rgba(255, 255, 255, 0.2);
+    border-radius: 4px;
+    cursor: pointer;
+    margin-top: 10px;
+    position: relative;
+    overflow: hidden;
+}
+
+.video-time-container {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  margin-top: 10px;
+  user-select: none;
+}
+
+.time-label {
+  font-size: 0.9rem;
+  color: #fff;
+  min-width: 40px;
+  text-align: center;
+}
+
+.custom-progress-bar {
+  flex-grow: 1;
+  height: 8px;
+  background: rgba(255, 255, 255, 0.2);
+  border-radius: 4px;
+  cursor: pointer;
+  position: relative;
+  overflow: hidden;
+}
+
+.progress-filled {
+  height: 100%;
+  background: #FD8804;
+  transition: width 0.1s linear;
+}
+
+.progress-thumb {
+  position: absolute;
+  top: 50%;
+  transform: translate(-50%, -50%);
+  width: 14px;
+  height: 14px;
+  background: white;
+  border-radius: 50%;
+  box-shadow: 0 0 4px rgba(0,0,0,0.5);
+  pointer-events: none;
+}
+
+.custom-video-controls.mute-controls {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 32px;
+  height: 32px;
+  background: rgba(0, 0, 0, 0.5);
+  border-radius: 50%;
+  cursor: pointer;
+  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.3);
+  transition: transform 0.3s ease, opacity 0.3s ease;
+}
+
+.custom-video-controls.mute-controls:hover {
+  transform: scale(1.1);
+}
+
+.custom-video-controls.mute-controls svg {
+  pointer-events: none;
+  width: 24px; 
+  height: 24px;
+}
+
+.mute-icon,
+.unmute-icon {
+  opacity: 1;
+  visibility: visible;
+  transition: opacity 0.3s ease, visibility 0.3s ease;
 }
